# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Attendance System API
  version: 0.1.0
  description: |
    Multi-tenant API scoped by `company_id`.
    When authentication is enabled, tenant is derived from credentials (for example API key context).
    Request-scoped `company_id` parameters are primarily for development/testing and non-auth deployments.
servers:
  - url: http://localhost:3000
...
paths:
...
  /api/employees-registry/{person_id}:
    get:
      tags: [employees-registry]
      operationId: getEmployee
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: person_id
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/company_id"
        - $ref: "#/components/parameters/x-company-id"
      responses:
        "200":
          description: Employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Employee not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    put:
      tags: [employees-registry]
      operationId: upsertEmployee
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: person_id
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/company_id"
        - $ref: "#/components/parameters/x-company-id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeUpsertRequest"
      responses:
        "200":
          description: Upserted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /api/attendance:
    get:
      tags: [attendance]
      operationId: getAttendance
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Attendance records
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    AgentBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
  parameters:
    company_id:
      name: company_id
      in: query
      description: |
        Tenant scoping. In production, prefer auth-derived tenant.
        Query param is primarily for dev/testing and non-auth deployments.
        CSV row values are not authoritative for tenant scoping.
      schema:
        type: string
        default: DEFAULT
    x-company-id:
      name: x-company-id
      in: header
      schema:
        type: string
...
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    EmployeeUpsertRequest:
      type: object
      properties:
        employee_code:
          type: string
        full_name:
          type: string
        default_rule_set_id:
          type: string
          format: uuid
        active:
          type: boolean
        metadata:
          type: object
...
    Employee:
      type: object
      properties:
        company_id:
          type: string
        person_id:
          type: string
        employee_code:
          type: string
        full_name:
          type: string
        default_rule_set_id:
          type: string
        active:
          type: boolean
        metadata:
          type: object
        created_at:
          type: string
        updated_at:
          type: string
...
    ErrorEnvelope:
      type: object
      required:
        - error
        - code
        - message
      properties:
        error:
          type: string
        code:
          type: string
        message:
          type: string
        details: {}